+++ HR.LeaveManagement.Application/Features/LeaveTypes/Handlers/Commands/DeleteLeaveTypeCommandHandler.cs
@@ -1,6 +1,9 @@
 ﻿using AutoMapper;
+using HR.LeaveManagement.Application.Contracts.Persistence;
+using HR.LeaveManagement.Application.Exceptions;
 using HR.LeaveManagement.Application.Features.LeaveTypes.Requests.Commands;
-using HR.LeaveManagement.Application.Persistence.Contracts;
+using HR.LeaveManagement.Application.Responses;
+using HR.LeaveManagement.Domain;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -13,7 +16,13 @@ namespace HR.LeaveManagement.Application.Features.LeaveTypes.Handlers.Commands
         public async Task Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
         {
             var leaveType = await leaveTypeRepository.GetAsync(request.Id);
-            await leaveTypeRepository.DeleteAsync(leaveType);
+
+            //if it is null, then it means that the record is not found in the database.
+            //Delete is idempotent, so we can just return without throwing an exception.
+            if (leaveType is not null)
+            {
+                await leaveTypeRepository.DeleteAsync(leaveType);
+            }
         }
     }
 }
